requirements.txt
flask
torch
torchvision
numpy
matplotlib
pillow
